requirements.txt
pygame==2.6.1
PyOpenGL==3.1.10
PyOpenGL_accelerate==3.1.10
